Move actions unit_tests to the correct folder (#794)
* Move actions unit_tests to the correct folder

The unit_tests for system_checks and pre_ponr_changes were in the
top-level `actions` folder, which is not where they belong anymore.


* Move EFIBootInfoMocked to be under unit_tests/__init__

This mock class is being used in both
unit_tests/actions/system_checks/efi_test.py and
unit_tests/grub_test.py, since it is a common and shared code, we are
moving it to be under unit_tests/__init__.py.


* Move backup_system unit_tests to the correct folder


---------
