+++ convert2rhel/unit_tests/__init__.py
@@ -22,6 +22,7 @@ from functools import wraps
 
 import pytest
 
+from convert2rhel import grub
 from convert2rhel.actions import STATUS_CODE
 from convert2rhel.utils import run_subprocess
 
@@ -375,3 +376,48 @@ def assert_actions_result(instance, status=_NO_USER_VALUE, error_id=_NO_USER_VAL
 
     if message != _NO_USER_VALUE:
         assert message in instance.message
+
+
+class EFIBootInfoMocked:
+    def __init__(
+        self,
+        current_bootnum="0001",
+        next_boot=None,
+        boot_order=("0001", "0002"),
+        entries=None,
+        exception=None,
+    ):
+        self.current_bootnum = current_bootnum
+        self.next_boot = next_boot
+        self.boot_order = boot_order
+        self.entries = entries
+        self.set_default_efi_entries()
+        self._exception = exception
+
+    def __call__(self):
+        """Tested functions call existing object instead of creating one.
+        The object is expected to be instantiated already when mocking
+        so tested functions are not creating new object but are calling already
+        the created one. From the point of the tested code, the behaviour is
+        same now.
+        """
+        if not self._exception:
+            return self
+        raise self._exception  # pylint: disable=raising-bad-type
+
+    def set_default_efi_entries(self):
+        if not self.entries:
+            self.entries = {
+                "0001": grub.EFIBootLoader(
+                    boot_number="0001",
+                    label="Centos Linux",
+                    active=True,
+                    efi_bin_source=r"HD(1,GPT,28c77f6b-3cd0-4b22-985f-c99903835d79,0x800,0x12c000)/File(\EFI\centos\shimx64.efi)",
+                ),
+                "0002": grub.EFIBootLoader(
+                    boot_number="0002",
+                    label="Foo label",
+                    active=True,
+                    efi_bin_source="FvVol(7cb8bdc9-f8eb-4f34-aaea-3ee4af6516a1)/FvFile(462caa21-7614-4503-836e-8ab6f4662331)",
+                ),
+            }
