+++ convert2rhel/unit_tests/grub_test.py
@@ -24,8 +24,7 @@ import pytest
 import six
 
 from convert2rhel import grub, utils
-from convert2rhel.unit_tests import run_subprocess_side_effect
-from convert2rhel.unit_tests.actions.efi_test import EFIBootInfoMocked
+from convert2rhel.unit_tests import EFIBootInfoMocked, run_subprocess_side_effect
 
 
 six.add_move(six.MovedModule("mock", "mock", "unittest.mock"))
